requirements.txt
environs==10.3.0
python-telegram-bot==13.15
google-cloud-dialogflow==2.27.0
